pyproject.toml
[project]
name = "mozhi-ai"
version = "0.1.0"
description = "Language AI / NLP research project"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "datasets>=3.0",
    "huggingface-hub>=0.25",
    "wikipedia-api>=0.7",
    "fasttext-wheel>=0.9",
    "indic-nlp-library>=0.92",
    "datasketch>=1.6",
    "pyarrow>=17.0",
]

[project.scripts]
mozhi = "mozhi.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mozhi"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]

[tool.ruff.lint.isort]
known-first-party = ["mozhi"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
